requirements.txt
fastapi[all]
pytest
sqlalchemy
psycopg2-binary
PyJWT